+++ talpid-core/src/split_tunnel/mod.rs
@@ -1,5 +1,9 @@
-#[cfg(target_os = "linux")]
-#[path = "linux.rs"]
+#[cfg(all(target_os = "linux", not(feature = "cgroups_v2")))]
+#[path = "linux_v1.rs"]
+mod imp;
+
+#[cfg(all(target_os = "linux", feature = "cgroups_v2"))]
+#[path = "linux_v2.rs"]
 mod imp;
 
 #[cfg(windows)]