+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Line wrap the file at 100 chars.                                              Th
 * **Security**: in case of vulnerabilities.
 
 ## [Unreleased]
+### Changed
+#### Linux
+- Upgrade split-tunneling to use cgroups v2, instead of the deprecated cgroups v1.
 
 
 ## [2025.14-beta2] - 2025-11-26